104-task

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,26 +7,26 @@
 int main(void)
 {
 	int x;
-	double prev = 1;
-	double curr = 2;
+	float prev = 1;
+	float curr = 2;
 
 	for (x = 1; x <= 98; x++)
 	{
 		if (x == 1)
 		{
-			printf("%.0lf", prev);
+			printf("%.0f", prev);
 			continue;
 		}
 		if (x == 2)
 		{
-			printf(", %.0lf", curr);
+			printf(", %.0f", curr);
 			continue;
 		}
 
 		curr += prev;
 		prev = curr - prev;
 
-		printf(", %.0lf", curr);
+		printf(", %.0f", curr);
 	}
 	printf("\n");
